refactor(form-builder): move temporary code away ObjectFieldSet
