+++ packages/@sanity/form-builder/src/inputs/ObjectInput/ObjectFieldSet.tsx
@@ -1,20 +1,11 @@
 // Render a fieldset inside the object input
-import React, {ForwardedRef, forwardRef, useMemo, useRef} from 'react'
+import React, {ForwardedRef, forwardRef, useMemo} from 'react'
 import {FormFieldPresence} from '@sanity/base/presence'
 import {FormFieldSet, FormFieldSetProps} from '@sanity/base/components'
-import {
-  Marker,
-  MultiFieldSet,
-  Path,
-  ConditionalPropertyCallbackContext,
-  ConditionalPropertyCallback,
-  SanityDocument,
-  CurrentUser,
-} from '@sanity/types'
-import {useCurrentUser} from '@sanity/base/hooks'
+import {Marker, MultiFieldSet, Path} from '@sanity/types'
 import {EMPTY_ARRAY} from '../../utils/empty'
-import withDocument from '../../utils/withDocument'
 import {getCollapsedWithDefaults} from './utils'
+import {ConditionalFieldWithDocument} from './Temporary_ConditionalFieldWithDocument'
 
 interface Props extends Omit<FormFieldSetProps, 'onFocus'> {
   fieldset: MultiFieldSet
@@ -95,109 +86,45 @@ export const ObjectFieldSet = forwardRef(function ObjectFieldSet(
     }
   }, [fieldNames, focusPath])
 
-  function isThenable(value: any) {
-    return typeof value?.then === 'function'
+  const renderFieldSet = () => {
+    return (
+      <FormFieldSet
+        {...rest}
+        key={fieldset.name}
+        title={fieldset.title}
+        description={fieldset.description}
+        level={level + 1}
+        columns={columns}
+        collapsible={collapsibleOpts.collapsible}
+        collapsed={isCollapsed}
+        onToggle={handleToggleFieldset}
+        __unstable_presence={isCollapsed ? childPresence : EMPTY_ARRAY}
+        __unstable_changeIndicator={false}
+        __unstable_markers={childMarkers}
+        ref={isCollapsed ? forwardedRef : null}
+      >
+        {children}
+      </FormFieldSet>
+    )
   }
-
-  function omitDeprecatedRole(user: CurrentUser): Omit<CurrentUser, 'role'> {
-    const {role, ...propsA} = user
-    return propsA
-  }
-
-  function useCheckCondition(
-    hidden: ConditionalPropertyCallback,
-    {document, currentUser, value}: ConditionalPropertyCallbackContext
-  ) {
-    const didWarn = useRef(false)
-    return useMemo(() => {
-      let result = false
-      try {
-        result = hidden({
-          document,
-          currentUser,
-          value,
-        })
-      } catch (err) {
-        console.error(`An error occurred while checking if field should be hidden: ${err.message}`)
-        return false
-      }
-      if (isThenable(result) && !didWarn.current) {
-        console.warn(
-          'The hidden option is either a promise or a promise returning function. Async callbacks for `hidden` option is not currently supported.'
-        )
-        return false
-      }
-      return result
-    }, [hidden, document, value, currentUser])
-  }
-
-  const ConditionalFieldWithDocument = withDocument(
-    forwardRef(function ConditionalFieldWithDocument(
-      propsB: {document: SanityDocument; value: unknown; hidden: ConditionalPropertyCallback},
-      ref /* ignore ref as there's no place to put it */
-    ) {
-      const {document, value, hidden} = propsB
-
-      const {value: currentUser} = useCurrentUser()
-      const shouldHide = useCheckCondition(hidden, {
-        currentUser: omitDeprecatedRole(currentUser),
-        document,
-        value,
-      })
-
-      return (
-        <>
-          {shouldHide ? null : (
-            <FormFieldSet
-              {...rest}
-              key={fieldset.name}
-              title={fieldset.title}
-              description={fieldset.description}
-              level={level + 1}
-              columns={columns}
-              collapsible={collapsibleOpts.collapsible}
-              collapsed={isCollapsed}
-              onToggle={handleToggleFieldset}
-              __unstable_presence={isCollapsed ? childPresence : EMPTY_ARRAY}
-              __unstable_changeIndicator={false}
-              __unstable_markers={childMarkers}
-              ref={isCollapsed ? forwardedRef : null}
-            >
-              {children}
-            </FormFieldSet>
-          )}
-        </>
-      )
-    })
-  )
 
   if (typeof fieldset.hidden === 'function') {
-    const object = {}
+    const fieldSetValuesObject = {}
     fieldset.fields.forEach((field) => {
-      object[field.name] = fieldValues[field.name]
+      fieldSetValuesObject[field.name] = fieldValues[field.name]
     })
 
-    return <ConditionalFieldWithDocument {...fieldset} value={object} hidden={fieldset.hidden} />
+    return (
+      <ConditionalFieldWithDocument
+        {...fieldset}
+        value={fieldSetValuesObject}
+        hidden={fieldset.hidden}
+      >
+        {renderFieldSet()}
+      </ConditionalFieldWithDocument>
+    )
   } else if (fieldset.hidden === true) {
     return null
   }
-  return (
-    <FormFieldSet
-      {...rest}
-      key={fieldset.name}
-      title={fieldset.title}
-      description={fieldset.description}
-      level={level + 1}
-      columns={columns}
-      collapsible={collapsibleOpts.collapsible}
-      collapsed={isCollapsed}
-      onToggle={handleToggleFieldset}
-      __unstable_presence={isCollapsed ? childPresence : EMPTY_ARRAY}
-      __unstable_changeIndicator={false}
-      __unstable_markers={childMarkers}
-      ref={isCollapsed ? forwardedRef : null}
-    >
-      {children}
-    </FormFieldSet>
-  )
+  return renderFieldSet()
 })
